Simple refactoring by creating spec.ts.
